Use Uv::venv() wherever we create venvs with uv (#768)

+++ rye/src/sync.rs
@@ -20,7 +20,7 @@ use crate::platform::get_toolchain_python_bin;
 use crate::pyproject::{read_venv_marker, write_venv_marker, ExpandedSources, PyProject};
 use crate::sources::py::PythonVersion;
 use crate::utils::{
-    get_venv_python_bin, mark_path_sync_ignore, set_proxy_variables, symlink_dir, CommandOutput,
+    get_venv_python_bin, set_proxy_variables, symlink_dir, update_venv_sync_marker, CommandOutput,
     IoPathContext,
 };
 use crate::uv::Uv;
@@ -334,19 +334,14 @@ pub fn create_virtualenv(
 ) -> Result<(), Error> {
     let py_bin = get_toolchain_python_bin(py_ver)?;
 
-    let mut venv_cmd = if Config::current().use_uv() {
+    if Config::current().use_uv() {
         // try to kill the empty venv if there is one as uv can't work otherwise.
         fs::remove_dir(venv).ok();
-        let mut venv_cmd = Uv::ensure_exists(output)?.cmd();
-        venv_cmd.arg("venv");
-        if output == CommandOutput::Verbose {
-            venv_cmd.arg("--verbose");
-        } else {
-            venv_cmd.arg("-q");
-        }
-        venv_cmd.arg("-p");
-        venv_cmd.arg(&py_bin);
-        venv_cmd
+        let uv = Uv::ensure_exists(output.quieter())?
+            .venv(venv, &py_bin, py_ver)
+            .context("failed to initialize virtualenv")?;
+        uv.write_marker()?;
+        uv.sync_marker();
     } else {
         // create the venv folder first so we can manipulate some flags on it.
         fs::create_dir_all(venv).path_context(venv, "unable to create virtualenv folder")?;
@@ -364,24 +359,16 @@ pub fn create_virtualenv(
         venv_cmd.arg("--no-seed");
         venv_cmd.arg("--prompt");
         venv_cmd.arg(prompt);
-        venv_cmd
+        venv_cmd.arg("--").arg(venv);
+        let status = venv_cmd
+            .status()
+            .context("unable to invoke virtualenv command")?;
+        if !status.success() {
+            bail!("failed to initialize virtualenv");
+        }
+
+        write_venv_marker(venv, py_ver)?;
     };
-
-    venv_cmd.arg("--").arg(venv);
-
-    let status = venv_cmd
-        .status()
-        .context("unable to invoke virtualenv command")?;
-    if !status.success() {
-        bail!("failed to initialize virtualenv");
-    }
-
-    write_venv_marker(venv, py_ver)?;
-
-    // uv can only do it now
-    if Config::current().use_uv() {
-        update_venv_sync_marker(output, venv);
-    }
 
     // On UNIX systems Python is unable to find the tcl config that is placed
     // outside of the virtualenv.  It also sometimes is entirely unable to find
@@ -392,20 +379,6 @@ pub fn create_virtualenv(
     }
 
     Ok(())
-}
-
-/// Update the cloud synchronization marker for the given path
-/// based on the config flag.
-fn update_venv_sync_marker(output: CommandOutput, venv_path: &Path) {
-    if let Err(err) = mark_path_sync_ignore(venv_path, Config::current().venv_mark_sync_ignore()) {
-        if output != CommandOutput::Quiet && Config::current().venv_mark_sync_ignore() {
-            warn!(
-                "unable to mark virtualenv {} ignored for cloud sync: {}",
-                venv_path.display(),
-                err
-            );
-        }
-    }
 }
 
 #[cfg(unix)]
